+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 1.3.6 (in development)
 ----------------------
 
+* The new :doc:`/plugins/play` lets you start your desktop music player with
+  the songs that match a query. Thanks to David Hamp-Gonsalves.
+
 Fixes:
 
 * :doc:`/plugins/mbsync`: Fix a regression in 1.3.5 that broke the plugin